LayoutTests/svg/W3C-SVG-1.1/struct-dom-01-b-expected.txt

KCanvasContainer at (0,0) size 479x359
  KCanvasContainer at (39,149) size 51x51 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
    KCanvasItem {rect} at (39,149) size 51x51 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M40.00,150.00L90.00,150.00L90.00,200.00L40.00,200.00"]
  KCanvasItem {rect} at (0,0) size 479x359 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M1.00,1.00L479.00,1.00L479.00,359.00L1.00,359.00"]
